+++ client/httplib/mock/mock_test.go
@@ -26,7 +26,6 @@ import (
 )
 
 func TestStartMock(t *testing.T) {
-
 	// httplib.defaultSetting.FilterChains = []httplib.FilterChain{mockFilter.FilterChain}
 
 	stub := StartMock()
@@ -41,7 +40,6 @@ func TestStartMock(t *testing.T) {
 
 	assert.Equal(t, expectedErr, err)
 	assert.Equal(t, expectedResp, resp)
-
 }
 
 // TestStartMock_Isolation Test StartMock that